backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
websockets==11.0.3
pydantic==2.4.2
sqlalchemy==2.0.22
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pydantic-settings==2.0.3
bcrypt==4.0.1
sqlalchemy-utils==0.41.1
python-dotenv==1.0.0
slowapi==0.1.8
email-validator==2.2.0
cryptography>=3.4.0
ecdsa>=0.13
rsa>=4.0
dnspython>=2.0.0
# Gerekebilecek diğer bağımlılıklar
limits>=2.3
